requirements.txt
pytest==9.0.2
selenium==4.39
webdriver-manager==4.0.2
pytest-html==4.1.1
pytest-xdist==2.5.0
python-dotenv==1.2.1